Usability improvements
Added nsis_info, debian_info, create_linux_setup_script, and archive parameters to AgkBuild to call the corresponding methods during the call without having to deal with calling them in the agkbuild file.

APK files are detected and the user is asked whether to overwrite or increment the build version.